[BE-156] Fixed controller unit-tests
