+++ pallets/minterest-model/src/mock.rs
@@ -1,1 +1,4 @@
 #![cfg(test)]
+
+pub type TestMinterestModel = minterest_model::Module<Runtime>;
+pub const BLOCKS_PER_YEAR: u128 = 5_256_000;
