+++ pallets/minterest-protocol/Cargo.toml
@@ -31,6 +31,7 @@ liquidity-pools = { path = "../liquidity-pools" }
 oracle = { path = "../oracle" }
 controller = { path = "../controller" }
 accounts = { path = "../accounts" }
+minterest-model = { path = "../minterest-model" }
 
 [features]
 default = ['std']
